Connect the internal data tree to the shared data structure
The parser generates an internal representation how the update should
run and store it into a struct swupdate_cfg. This connects the shared
object "installer" to the internal database to retrieve information
about the update via IPC.

It is thought that the database is for IPC just read-only and no direct
access is allowed.

Signed-off-by: Stefano Babic <sbabic@denx.de>