Accept selections from IPC if configured at startup
If selection can be set via IPC, it could be possible to activate a
selection that can damage the board. For example, the IPC selects to
install into the running rootfs instead of the stand-by copy.

Do not accept any incoming selection, but verify it with a list that is
configured at the startup via command line parameter. The list is set
using --accepted-select, that can be issued multiple times.

The behavior is compatible with the past because if no --accepted-select
is issued, the feature is completely disabled and no selection can be
set via IPC.

Signed-off-by: Stefano Babic <sbabic@denx.de>